Centralize element message notification in model base class
Move element notification logic from individual model implementations to a
common base class method to reduce code duplication and improve maintainability.

Changes:
- Moved meshXElementIF definition from meshx_element_class.hpp to
  meshx_fwd_decl.hpp for better forward declaration support
- Added prepare_element_msg() pure virtual method to meshXModelIF
- Added send_to_parent_element() common implementation to handle element
  state change and notification in one place
- Added update_element_state_change_header() virtual method for type-specific
  header casting in client/server models
- Refactored all model implementations (Config, Generic, Light) to store
  element messages as member variables and implement prepare_element_msg()
- Removed direct on_model_cb() calls from individual model implementations

This eliminates repetitive boilerplate code across 20+ model classes and
establishes a consistent pattern for element-state-to-model communication.

+++ main/component/meshx/elements/src/variants/meshx_root_element.cpp
@@ -47,7 +47,7 @@ uint8_t meshXRootElement :: list_sig_models()
     auto config_model = std::make_unique<meshXConfigModel>(this);
     this->get_sig_models().push_back(std::move(config_model));
 
-    return this->get_sig_models().size();
+    return (uint8_t)this->get_sig_models().size();
 }
 
 /**
@@ -87,7 +87,6 @@ meshx_err_t meshXRootElement :: element_state_change_notify(meshx_ptr_t param, s
     MESHX_LOGI(MODULE_ID_ELEMENT_ROOT,
         "Root element received model callback");
 
-    // TODO: Handle config model specific events if needed
     // For now, just log the event and return success
 
     return MESHX_SUCCESS;
